Clone the internal state rather than pointing to it
The standard use for getting the RNG state is to take a snapshot of the
status at a given time, not to point to its evolving state for future referal.
Besides, since the actual state also comprises of two integers that
cannot be easily kept up-to-date (unless wrapping them in a Tensor), it
is more coherent to ensure that the Tensor is not kept up-to-date
either.