Use proper median-of-three partitioning
In #176, we simply swapped middle element and left element
instead of using the proper median-of-three advocated by Sedgewick.
There is actually a sound reason for median-of-three, in spite of the
slightly added cost: it balances the partitions better, therefore
speeding things up on average.
Crude benchmarking on uniformly random array does not show much
of a difference (or a slight 10% performance cost, if any), but I will
sleep better knowing that I stuck to the proper way. The earlier
crude hack could have biten us in the back on some inputs -- or
Sedgweick would have done it instead of Median-of-three.