Improve benchmark
Add randomization, nicer plots, and take ratio of means instead of
mean of ratios: it doesn't change
anything here because the variance of the runtimes is very small
compared to its mean. It would, however, make a change if we
were taking ratio of gaussians with very small means and high
variance.